refactor(react): rename withReconfigure to withReconfiguration
