+++ packages/react/modules/components/configure-adapter/index.js
@@ -1,3 +1,3 @@
 export default from './configure-adapter';
 export { AdapterContext } from './configure-adapter';
-export { default as withReconfigure } from './with-reconfigure';
+export { default as withReconfiguration } from './with-reconfiguration';
